+++ clease/datastructures/four_vector.py
@@ -12,24 +12,20 @@ import attr
 __all__ = ("FourVector", "construct_four_vectors")
 
 
-def _int_converter(x):
-    """Allow for NumPy integer types (convert to regular int)"""
-    if isinstance(x, np.integer):
-        return int(x)
-    return x
+_FV_KWARGS = dict(validator=attr.validators.instance_of((np.integer, int)))
 
 
 @attr.s(frozen=True, order=True)
 class FourVector:
     """Container for a vector in 4-vector space, i.e. a vector of [ix, iy, iz, sublattice].
-    Represents the position of a site in terms of the number of repitions of the primitive atoms,
+    Represents the position of a site in terms of the number of repetitions of the primitive atoms,
     as well as which sublattice it belongs to in that cell.
     """
 
-    ix: int = attr.ib(converter=_int_converter, validator=attr.validators.instance_of(int))
-    iy: int = attr.ib(converter=_int_converter, validator=attr.validators.instance_of(int))
-    iz: int = attr.ib(converter=_int_converter, validator=attr.validators.instance_of(int))
-    sublattice: int = attr.ib(converter=_int_converter, validator=attr.validators.instance_of(int))
+    ix: int = attr.ib(**_FV_KWARGS)
+    iy: int = attr.ib(**_FV_KWARGS)
+    iz: int = attr.ib(**_FV_KWARGS)
+    sublattice: int = attr.ib(**_FV_KWARGS)
 
     def to_cartesian(self, prim: Atoms, transposed_cell: np.ndarray = None) -> np.ndarray:
         """Convert the four vector into cartesian coordinates
@@ -87,11 +83,36 @@ class FourVector:
         FourVector(ix=1, iy=0, iz=0, sublattice=1)
         """
         if not isinstance(other, FourVector):
-            raise NotImplementedError(f"Shift must be by another FourVector, got {type(other)}")
+            raise NotImplementedError(f"Shift must be by another FourVector, got {other!r}")
 
         return FourVector(
             self.ix + other.ix, self.iy + other.iy, self.iz + other.iz, self.sublattice
         )
+
+    def shift_xyz_and_modulo(self, other: "FourVector", nx: int, ny: int, nz: int) -> "FourVector":
+        """
+        Shift the four vector, similarly to "shift_xyz". Additionally, apply a modulo
+        of (nx, ny, nz) to (ix, iy, iz).
+        Useful for wrapping a four-vector back into a supercell, which is a (nx, ny, nz)
+        repetition of the primitive. The sublattice remains the same.
+
+        Return a new FourVector instance.
+
+        Example:
+
+        >>> from clease.datastructures.four_vector import FourVector
+        >>> a = FourVector(-1, -1, 3, 0)
+        >>> b = FourVector(0, 0, 0, 0)
+        >>> a.shift_xyz_and_modulo(b, 2, 2, 1)
+        FourVector(ix=1, iy=1, iz=0, sublattice=0)
+        """
+        if not isinstance(other, FourVector):
+            raise NotImplementedError(f"Shift must be by another FourVector, got {other!r}")
+
+        ix = (self.ix + other.ix) % nx
+        iy = (self.iy + other.iy) % ny
+        iz = (self.iz + other.iz) % nz
+        return FourVector(ix, iy, iz, self.sublattice)
 
 
 class _Box(NamedTuple):